refactor(core): separate Core function into different files for better maintainability
